LayoutTests/http/wpt/webaudio/the-audio-api/the-audioworklet-interface/dynamic-import-is-prohibited.https-expected.txt


PASS dynamic-import is prohibited in AudioWorklets
